requirements.txt
accelerate==0.21.0
diffusers==0.19.3
gradio==3.40.1
invisible-watermark==0.2.0
Pillow==10.0.0
torch==2.0.1
transformers==4.31.0
